Split the tests up into separate files
