+++ packages/postgraphile-core/__tests__/integration/schema/defaultOptions.subscriptions.test.js
@@ -1,0 +1,8 @@
+const core = require("./core");
+
+test(
+  "should be same with subscriptions option set",
+  core.test(__filename, ["a", "b", "c"], {
+    subscriptions: true,
+  })
+);
