+++ packages/postgraphile-core/__tests__/integration/schema/indexes.index_expressions.test.js
@@ -1,0 +1,8 @@
+const core = require("./core");
+
+test(
+  "prints a schema for index_expressions",
+  core.test(__filename, ["index_expressions"], {
+    disableDefaultMutations: true,
+  })
+);
