+++ packages/postgraphile-core/__tests__/integration/schema/indexes.index_type_checker_plugin.test.js
@@ -1,0 +1,20 @@
+const core = require("./core");
+
+const IndexTypeCheckerPlugin = builder => {
+  builder.hook("build", build => {
+    const { pgIntrospectionResultsByKind } = build;
+    if (
+      !pgIntrospectionResultsByKind.index.every(idx => idx.indexType != null)
+    ) {
+      throw new Error("indexType missing");
+    }
+    return build;
+  });
+};
+
+test(
+  "index types are present on introspection",
+  core.test(__filename, ["index_expressions"], {
+    appendPlugins: [IndexTypeCheckerPlugin],
+  })
+);
