+++ packages/postgraphile-core/__tests__/integration/schema/indexes.test.js
@@ -6,29 +6,3 @@ test(
     ignoreIndexes: false,
   })
 );
-
-test(
-  "prints a schema for index_expressions",
-  core.test(__filename, ["index_expressions"], {
-    disableDefaultMutations: true,
-  })
-);
-
-const IndexTypeCheckerPlugin = builder => {
-  builder.hook("build", build => {
-    const { pgIntrospectionResultsByKind } = build;
-    if (
-      !pgIntrospectionResultsByKind.index.every(idx => idx.indexType != null)
-    ) {
-      throw new Error("indexType missing");
-    }
-    return build;
-  });
-};
-
-test(
-  "index types are present on introspection",
-  core.test(__filename, ["index_expressions"], {
-    appendPlugins: [IndexTypeCheckerPlugin],
-  })
-);
