+++ packages/postgraphile-core/__tests__/integration/schema/network_types.custom.test.js
@@ -1,10 +1,10 @@
 const core = require("./core");
 
 test(
-  "prints a schema to test network scalars without using custom network scalar types",
+  "prints a schema to test network scalars using custom network scalar types",
   core.test(__filename, ["network_types"], {
     graphileBuildOptions: {
-      pgUseCustomNetworkScalars: false,
+      pgUseCustomNetworkScalars: true,
     },
   })
 );
