+++ packages/postgraphile-core/__tests__/integration/schema/omit-rename.omitcolumns.execute.test.js
@@ -1,0 +1,13 @@
+const core = require("./core");
+
+test(
+  "omit execute on computed column",
+  core.test(
+    __filename,
+    ["d"],
+    {},
+    `
+comment on function d.person_full_name(d.person) is E'@omit execute';
+`
+  )
+);
