+++ packages/postgraphile-core/__tests__/integration/schema/omit-rename.omitcolumns.shows-title-asterisk.test.js
@@ -1,0 +1,14 @@
+const core = require("./core");
+
+test(
+  "omit on column",
+  core.test(
+    __filename,
+    ["d"],
+    {},
+    `
+comment on column d.tv_shows.title is E'@omit *';
+comment on column d.tv_episodes.title is E'@omit';
+`
+  )
+);
