+++ packages/postgraphile-core/__tests__/integration/schema/omit-rename.omitcolumns.title-order.test.js
@@ -1,13 +1,13 @@
 const core = require("./core");
 
 test(
-  "omit create on column",
+  "omit order on column",
   core.test(
     __filename,
     ["d"],
     {},
     `
-comment on column d.tv_shows.title is E'@omit create';
+comment on column d.tv_episodes.title is E'@omit order';
 `
   )
 );
