+++ packages/postgraphile-core/__tests__/integration/schema/omit-rename.omitcolumns.update-title.test.js
@@ -1,13 +1,13 @@
 const core = require("./core");
 
 test(
-  "omit create on column",
+  "omit update on column",
   core.test(
     __filename,
     ["d"],
     {},
     `
-comment on column d.tv_shows.title is E'@omit create';
+comment on column d.tv_shows.title is E'@omit update';
 `
   )
 );
