+++ packages/postgraphile-core/__tests__/integration/schema/omit-rename.omitstuff.films-asterisk.test.js
@@ -1,0 +1,14 @@
+const core = require("./core");
+
+test(
+  "omit and omit everything on table",
+  core.test(
+    __filename,
+    ["d"],
+    {},
+    `
+comment on table d.films is E'@omit *';
+comment on table d.tv_shows is E'@omit';
+`
+  )
+);
