+++ packages/postgraphile-core/__tests__/integration/schema/pg11.test.js
@@ -4,11 +4,3 @@ test(
   "prints a schema to test PG11-specific features with default options",
   core.test(__filename, ["pg11"])
 );
-test(
-  "prints a schema to test PG11-specific features with custom network scalars",
-  core.test(__filename, ["pg11"], {
-    graphileBuildOptions: {
-      pgUseCustomNetworkScalars: true,
-    },
-  })
-);
