+++ packages/postgraphile-core/__tests__/integration/schema/rbac.ignore.test.js
@@ -1,0 +1,8 @@
+const core = require("./core");
+
+test(
+  "prints a schema from non-root role, with RBAC ignored",
+  core.test(__filename, ["a", "b", "c"], { ignoreRBAC: true }, client =>
+    client.query("set role postgraphile_test_authenticator")
+  )
+);
