+++ packages/postgraphile-core/__tests__/integration/schema/rbac.test.js
@@ -6,9 +6,3 @@ test(
     client.query("set role postgraphile_test_authenticator")
   )
 );
-test(
-  "prints a schema from non-root role, with RBAC ignored",
-  core.test(__filename, ["a", "b", "c"], { ignoreRBAC: true }, client =>
-    client.query("set role postgraphile_test_authenticator")
-  )
-);
