+++ packages/postgraphile-core/__tests__/integration/schema/simple-collections.pets-simple.test.js
@@ -1,0 +1,17 @@
+const core = require("./core");
+
+test(
+  "simple collection for relation",
+  core.test(
+    __filename,
+    "simple_collections",
+    {
+      simpleCollections: "omit",
+      setofFunctionsContainNulls: false,
+    },
+    pgClient =>
+      pgClient.query(`
+        comment on constraint pets_owner_id_fkey on simple_collections.pets is E'@simpleCollections only';
+      `)
+  )
+);
